+++ src/Encryptor/Aes256FixedEncryptor.php
@@ -11,6 +11,11 @@ namespace PrecisionSoft\Doctrine\Encrypt\Encryptor;
 use PrecisionSoft\Doctrine\Encrypt\Exception\Exception;
 use PrecisionSoft\Doctrine\Encrypt\Type\Aes256FixedType;
 
+/**
+ * BREAKING CHANGE: nonce derivation now uses an HKDF-derived nonceKey instead of the raw salt.
+ * Existing data encrypted with a previous version will NOT be decryptable.
+ * Use the database decrypt/encrypt commands to migrate existing data before upgrading.
+ */
 class Aes256FixedEncryptor extends AbstractEncryptor
 {
     public function getTypeClass(): string
@@ -26,7 +31,7 @@ class Aes256FixedEncryptor extends AbstractEncryptor
             throw new Exception(\sprintf('failed to get IV length for cipher "%s"', static::ALGORITHM));
         }
 
-        $hash = \hash_hmac('sha256', $data, $this->salt, true);
+        $hash = \hash_hmac('sha256', $data, $this->nonceKey, true);
 
         return \substr($hash, 0, $ivLength);
     }