+++ tests/PrecisionSoftDoctrineEncryptBundleTest.php
@@ -10,8 +10,6 @@ namespace PrecisionSoft\Doctrine\Encrypt\Test;
 
 use Doctrine\DBAL\Types\Type;
 use Mockery;
-use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
-use PHPUnit\Framework\TestCase;
 use PrecisionSoft\Doctrine\Encrypt\Contract\EncryptorInterface;
 use PrecisionSoft\Doctrine\Encrypt\Encryptor\Aes256Encryptor;
 use PrecisionSoft\Doctrine\Encrypt\Encryptor\Aes256FixedEncryptor;
@@ -21,17 +19,24 @@ use PrecisionSoft\Doctrine\Encrypt\PrecisionSoftDoctrineEncryptBundle;
 use PrecisionSoft\Doctrine\Encrypt\Service\EncryptorFactory;
 use PrecisionSoft\Doctrine\Encrypt\Type\Aes256FixedType;
 use PrecisionSoft\Doctrine\Encrypt\Type\Aes256Type;
+use PrecisionSoft\Symfony\Phpunit\MockDto;
+use PrecisionSoft\Symfony\Phpunit\TestCase\AbstractTestCase;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * @internal
  */
-final class PrecisionSoftDoctrineEncryptBundleTest extends TestCase
+final class PrecisionSoftDoctrineEncryptBundleTest extends AbstractTestCase
 {
-    use MockeryPHPUnitIntegration;
+    public static function getMockDto(): MockDto
+    {
+        return new MockDto(PrecisionSoftDoctrineEncryptBundle::class);
+    }
 
     protected function setUp(): void
     {
+        parent::setUp();
+
         $this->removeTypeIfExists('encryptedAes256_bundle_test');
         $this->removeTypeIfExists('encryptedAes256fixed_bundle_test');
     }
@@ -145,7 +150,7 @@ final class PrecisionSoftDoctrineEncryptBundleTest extends TestCase
 
         $precisionSoftDoctrineEncryptBundle->boot();
 
-        static::assertSame(true, true);
+        $this->expectNotToPerformAssertions();
     }
 
     private function removeTypeIfExists(string $typeName): void